discord-bot/requirements.txt
discord.py>=2.3.0
boto3>=1.38.0
